Global clean-up.
Started the strategy of estimate the energy by classifying.

Modified the Magic inception with the new double-attention block.

Started to implement amoebanet (just created the file).

Gived a try to SE-IncV3, not really good beacuse i have to destroy the architecture in order to make it fit the not so big picture.

Gived a try to deep svaso but it is extremely inefficient.

Deleted Resnet and autokeras.